Biomedical RAG Assistant/requirements.txt
streamlit
langchain
langchain-chroma
langchain-huggingface
langchain-google-genai
sentence-transformers
chromadb
pytesseract
pdf2image
unstructured
python-dotenv
